+++ internal/adapters/https/pkg/app/app.go
@@ -110,10 +110,10 @@ func NewApp(
 
 func (p *sApp) Run(pCtx context.Context) error {
 	services := []internal_types.IServiceF{
-		p.runHTTPIntAdapter,
-		p.runHTTPExtAdapter,
-		p.runHTTPIntRelayer,
-		p.runHTTPExtRelayer,
+		p.runIntAdapter,
+		p.runExtAdapter,
+		p.runIntRelayer,
+		p.runExtRelayer,
 	}
 
 	ctx, cancel := context.WithCancel(pCtx)
@@ -171,7 +171,7 @@ func (p *sApp) disable(pCancel context.CancelFunc, pWg *sync.WaitGroup) state.IS
 	}
 }
 
-func (p *sApp) runHTTPIntAdapter(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
+func (p *sApp) runIntAdapter(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
 	defer wg.Done()
 
 	if err := p.fIntAdapter.Run(pCtx); err != nil {
@@ -180,7 +180,7 @@ func (p *sApp) runHTTPIntAdapter(pCtx context.Context, wg *sync.WaitGroup, pChEr
 	}
 }
 
-func (p *sApp) runHTTPExtAdapter(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
+func (p *sApp) runExtAdapter(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
 	defer wg.Done()
 
 	if err := p.fExtAdapter.Run(pCtx); err != nil {
@@ -189,7 +189,7 @@ func (p *sApp) runHTTPExtAdapter(pCtx context.Context, wg *sync.WaitGroup, pChEr
 	}
 }
 
-func (p *sApp) runHTTPIntRelayer(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
+func (p *sApp) runIntRelayer(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
 	defer wg.Done()
 
 	for {
@@ -211,7 +211,7 @@ func (p *sApp) runHTTPIntRelayer(pCtx context.Context, wg *sync.WaitGroup, pChEr
 	}
 }
 
-func (p *sApp) runHTTPExtRelayer(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
+func (p *sApp) runExtRelayer(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
 	defer wg.Done()
 
 	for {