+++ internal/adapters/tcp/pkg/app/app.go
@@ -103,10 +103,10 @@ func NewApp(pCfg config.IConfig, pPathTo string) types.IRunner {
 
 func (p *sApp) Run(pCtx context.Context) error {
 	services := []internal_types.IServiceF{
-		p.runTCPAdapter,
-		p.runTCPRelayer,
-		p.runHTTPAdapter,
-		p.runHTTPRelayer,
+		p.runExtAdapter,
+		p.runIntAdapter,
+		p.runExtRelayer,
+		p.runIntRelayer,
 	}
 
 	ctx, cancel := context.WithCancel(pCtx)
@@ -164,7 +164,7 @@ func (p *sApp) disable(pCancel context.CancelFunc, pWg *sync.WaitGroup) state.IS
 	}
 }
 
-func (p *sApp) runTCPAdapter(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
+func (p *sApp) runExtAdapter(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
 	defer wg.Done()
 
 	if err := p.fExtAdapter.Run(pCtx); err != nil {
@@ -173,7 +173,7 @@ func (p *sApp) runTCPAdapter(pCtx context.Context, wg *sync.WaitGroup, pChErr ch
 	}
 }
 
-func (p *sApp) runHTTPAdapter(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
+func (p *sApp) runIntAdapter(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
 	defer wg.Done()
 
 	if err := p.fIntAdapter.Run(pCtx); err != nil {
@@ -182,7 +182,7 @@ func (p *sApp) runHTTPAdapter(pCtx context.Context, wg *sync.WaitGroup, pChErr c
 	}
 }
 
-func (p *sApp) runTCPRelayer(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
+func (p *sApp) runExtRelayer(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
 	defer wg.Done()
 
 	for {
@@ -209,7 +209,7 @@ func (p *sApp) runTCPRelayer(pCtx context.Context, wg *sync.WaitGroup, pChErr ch
 	}
 }
 
-func (p *sApp) runHTTPRelayer(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
+func (p *sApp) runIntRelayer(pCtx context.Context, wg *sync.WaitGroup, pChErr chan<- error) {
 	defer wg.Done()
 
 	for {