+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 ### CHANGES
 
 - `cmd/hls/hls-messenger`: change API paths, LoadMessages -> LoadMessage, CountMessages -> GetChatSize
-- `cmd/hla/*`: add suffix `&` for internal adapter
+- `cmd/hla/*`: add suffix `<` for internal adapter
 - `cmd/hla/*`: log NOCON: WARN -> INFO
 
 <!-- ... -->